creado el componente donaciones, importado y agregado a la vista home
