+++ src/front/js/layout.js
@@ -16,7 +16,7 @@ import { PublicacionNueva } from "./pages/vistaPublicacion";
 import { RecuperarContraseña } from "./pages/vistaRecuperar";
 import { Encontrados } from "./pages/vista-encontrados";
 import { Adopcion } from "./pages/vista-adopcion";
-import { Donaciones } from "./pages/donaciones";
+// import { Donaciones } from "./pages/donaciones";
 
 // import { Login } from "./component/login";
 
@@ -50,7 +50,7 @@ const Layout = () => {
             <Route element={<Perdidos />} path="/perdidos" />
             <Route element={<Encontrados />} path="/encontrados" />
             <Route element={<Adopcion />} path="/en-adopcion" />
-            <Route element={<Donaciones />} path="/donaciones" />
+            {/* <Route element={<Donaciones />} path="/donaciones" /> */}
           </Routes>{" "}
           <FooterVariable />
         </ScrollToTop>{" "}
