+++ src/front/js/pages/home.js
@@ -10,6 +10,7 @@ import { Mascotas } from "../component/mascotas.jsx";
 // import { Footerhome } from "../component/footerhome.jsx";
 // import { Link } from "react-router-dom";
 import "../../styles/home.css";
+import { Donaciones } from "../component/donaciones.jsx";
 // import { element } from "prop-types";
 
 export const Home = () => {
@@ -243,6 +244,8 @@ export const Home = () => {
           </a>
         </div>
       </div>
+
+      <Donaciones />
     </div>
     //   </div>
     // </div>
